fp32 lm head

+++ vllm/model_executor/models/afmoe.py
@@ -685,7 +685,7 @@ class AfmoeForCausalLM(nn.Module, SupportsPP, SupportsLoRA):
         return hidden_states
 
     def compute_logits(self, hidden_states: torch.Tensor) -> torch.Tensor | None:
-        logits = self.logits_processor(self.lm_head, hidden_states)
+        logits = self.logits_processor(self.lm_head, hidden_states.float())
         return logits
 
     def load_weights(self, weights: Iterable[tuple[str, torch.Tensor]]) -> set[str]:
